pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=0.9.9",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.5.2",
    "langchain-google-genai>=2.1.7",
    "isort>=6.0.1",
    "black>=25.1.0",
]
